+++ app/cypress/support/commands.ts
@@ -13,15 +13,20 @@ Cypress.Commands.add('findBySel', { prevSubject: true }, (subject, selector, ...
 })
 
 Cypress.Commands.add('selectTokenByChain', (selector, direction, chain, token) => {
-  cy.getBySel(`${selector}-${direction}`).findBySel('chain-select-trigger').click()
+  cy.getBySel(`${selector}-${direction}`)
+    .findBySel('chain-select-trigger')
+    .contains('Chain')
+    .click()
   cy.contains(chain).click()
   cy.getBySel(`token-list-${direction}`).contains(token).click()
 })
 
 Cypress.Commands.add('connectWallet', () => {
+  expect(cy).property('initWallet').to.be.a('function')
+
   cy.initWallet([Alice], DAPP_NAME)
-  cy.contains('Connect').should('be.visible')
   cy.selectTokenByChain('chain-container', 'source', 'Asset Hub', 'DOT')
+  cy.contains('Connect').should('be.visible')
   cy.getBySel('chain-container-source').contains('Connect').click()
   cy.contains('Connect Wallet')
     .should('be.visible')
