+++ app/src/components/ChainTokenSelect.tsx
@@ -6,12 +6,8 @@ import { ManualAddressInput } from '@/models/select'
 import { Token } from '@/models/token'
 import { cn } from '@/utils/cn'
 import { reorderOptionsBySelectedItem } from '@/utils/sort'
-import NumberFlow from '@number-flow/react'
-import Image from 'next/image'
-import { ChangeEvent, ReactNode, RefObject, useMemo, useRef, useState } from 'react'
+import { ReactNode, RefObject, useMemo, useRef, useState } from 'react'
 import { twMerge } from 'tailwind-merge'
-import { colors } from '../../tailwind.config'
-import Button from './Button'
 import ChainTrigger from './ChainTrigger'
 import Dropdown from './Dropdown'
 import TokenAmountInput from '@/components/chainTokenSelect/TokenAmountInput'
@@ -227,239 +223,4 @@ interface TokenAmountInputProps {
   triggerRef: RefObject<HTMLDivElement | null>
   disabled?: boolean
   inDollars?: number
-}
-
-const TokenAmountInput = ({
-  token,
-  amount,
-  disabled,
-  onTriggerClick,
-  triggerRef,
-  inDollars,
-}: TokenAmountInputProps) => {
-  const showVerticalDivider = !!amount?.value || !!amount?.placeholder
-  const handleAmountChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const newVal = e.target.value === '' ? null : parseFloat(e.target.value)
-    amount?.onChange?.(newVal)
-  }
-
-  return (
-    <Tooltip content={amount?.error}>
-      <div
-        ref={triggerRef}
-        className={cn(
-          'flex items-center justify-between rounded-md rounded-t-none border-1 border-t-0 border-turtle-level3 bg-background px-3 text-sm',
-          !disabled && 'cursor-pointer',
-          disabled && 'opacity-30',
-          amount?.error && 'border-1 border-turtle-error',
-        )}
-      >
-        <div className="flex h-[3.5rem] flex-grow items-center gap-1">
-          <div
-            className="flex items-center gap-1"
-            data-cy="token-select-trigger"
-            onClick={disabled ? undefined : onTriggerClick}
-          >
-            {token.value ? (
-              <>
-                <TokenLogo
-                  token={token.value}
-                  sourceChain={token.sourceChainToDetermineOriginBanner}
-                />
-                <span className="ml-1 text-nowrap" data-cy="token-select-symbol">
-                  {token.value.symbol}
-                </span>
-              </>
-            ) : (
-              <>
-                <TokenIcon />
-                Token
-              </>
-            )}
-            <ChevronDown strokeWidth={0.2} className="ml-1" />
-            {showVerticalDivider && <VerticalDivider />}
-          </div>
-
-          <div className="align-center ml-1 flex flex-col">
-            <input
-              data-cy="amount-input"
-              disabled={disabled || amount?.disabled}
-              type="number"
-              className={cn(
-                'bg-transparent text-sm focus:border-0 focus:outline-none min-[350px]:text-base sm:text-xl',
-                inDollars && 'animate-slide-up-slight',
-                amount?.error && 'text-turtle-error',
-              )}
-              placeholder={amount?.placeholder ?? 'Amount'}
-              value={amount?.value ?? ''}
-              onChange={handleAmountChange}
-              onClick={e => e.stopPropagation()}
-              onWheel={e => e.target instanceof HTMLElement && e.target.blur()}
-              autoFocus
-            />
-            {inDollars && (
-              <div className={'animate-slide-up mt-[-3px] text-sm text-turtle-level4'}>
-                <NumberFlow
-                  value={Math.min(inDollars, maxDollars)}
-                  prefix="$"
-                  format={numberFlowFormat}
-                />
-              </div>
-            )}
-          </div>
-        </div>
-
-        {amount?.trailingAction && (
-          <div className="absolute right-0 ml-2 mr-3 bg-white">{amount.trailingAction}</div>
-        )}
-      </div>
-    </Tooltip>
-  )
-}
-
-interface ChainListProps {
-  searchString: string
-  setSearchString: (value: string) => void
-  options: Chain[]
-  selectedChain: Chain | null
-  clearable?: boolean
-  onSelect: (chain: Chain) => void
-  onClear: () => void
-}
-
-const ChainList = ({
-  searchString,
-  setSearchString,
-  options,
-  selectedChain,
-  clearable,
-  onSelect,
-  onClear,
-}: ChainListProps) => {
-  return (
-    <>
-      <SearchBar placeholder="Search" value={searchString} onChange={setSearchString} />
-      <div className="max-h-[15rem] overflow-y-auto">
-        <ul className="flex flex-col">
-          {options.map(option => (
-            <li
-              key={option.uid}
-              className={cn(
-                'flex cursor-pointer items-center justify-between px-3 py-3 hover:bg-turtle-level1',
-                selectedChain?.uid === option.uid &&
-                  'bg-turtle-secondary-light hover:bg-turtle-secondary-light',
-              )}
-              onClick={() => onSelect(option)}
-            >
-              <div className="flex items-center gap-2">
-                <Image
-                  src={option.logoURI}
-                  alt={option.name}
-                  width={24}
-                  height={24}
-                  priority
-                  className="h-[2rem] w-[2rem] rounded-full border-1 border-turtle-foreground bg-background"
-                />
-                <span className="text-sm">{option.name}</span>
-              </div>
-
-              {selectedChain?.uid === option.uid && clearable && <ClearButton onClick={onClear} />}
-            </li>
-          ))}
-        </ul>
-      </div>
-    </>
-  )
-}
-
-interface TokenListProps {
-  searchString: string
-  setSearchString: (value: string) => void
-  options: Token[]
-  selectedToken: Token | null
-  clearable?: boolean
-  sourceChainToDetermineOriginBanner: Chain | null
-  onSelect: (token: Token) => void
-  onClear: () => void
-}
-
-const TokenList = ({
-  searchString,
-  setSearchString,
-  options,
-  selectedToken,
-  clearable,
-  sourceChainToDetermineOriginBanner,
-  onSelect,
-  onClear,
-}: TokenListProps) => {
-  return (
-    <>
-      <SearchBar placeholder="Search" value={searchString} onChange={setSearchString} />
-      <div className="max-h-[15rem] overflow-y-auto">
-        <ul className="flex flex-col">
-          {options.map(option => (
-            <li
-              key={option.id}
-              className={cn(
-                'flex cursor-pointer items-center justify-between px-3 py-3 hover:bg-turtle-level1',
-                selectedToken?.id === option.id &&
-                  'bg-turtle-secondary-light hover:bg-turtle-secondary-light',
-              )}
-              onClick={() => onSelect(option)}
-            >
-              <div className="flex items-center gap-2">
-                <TokenLogo token={option} sourceChain={sourceChainToDetermineOriginBanner} />
-                <span className="text-sm">{option.symbol}</span>
-              </div>
-
-              {selectedToken?.id === option.id && clearable && <ClearButton onClick={onClear} />}
-            </li>
-          ))}
-        </ul>
-      </div>
-    </>
-  )
-}
-
-const ClearButton = ({ onClick }: { onClick: () => void }) => {
-  return (
-    <Button
-      label="Clear"
-      size="sm"
-      variant="outline"
-      className="z-10 h-[28px] w-[28px] min-w-5 border-turtle-secondary text-sm hover:border-turtle-secondary"
-      onClick={onClick}
-    >
-      <div className="flex items-center gap-1 text-turtle-foreground">
-        <Cross stroke={colors['turtle-secondary']} />
-      </div>
-    </Button>
-  )
-}
-
-const SearchBar = ({
-  placeholder,
-  value,
-  onChange,
-}: {
-  placeholder: string
-  value: string
-  onChange: (value: string) => void
-}) => {
-  return (
-    <div className="sticky top-0 z-20 flex items-center gap-2 border-b-1 border-turtle-level3 bg-turtle-background px-3 py-3">
-      <div className="flex h-[2rem] w-[2rem] shrink-0 items-center justify-center">
-        <SearchIcon fill={colors['turtle-level5']} width={17} height={17} />
-      </div>
-
-      <input
-        type="text"
-        value={value}
-        onChange={e => onChange(e.target.value)}
-        placeholder={placeholder}
-        className="w-full rounded-md border-0 bg-transparent text-sm placeholder:text-turtle-level5 focus:border-0 focus:outline-none"
-      />
-    </div>
-  )
 }
